--- knock/src/lib/hooks/useCopyToClipboard.ts
@@ -1,21 +1,0 @@
-import { useState } from 'react';
-
-const useCopyToClipboard = (): [boolean, (text: string) => void] => {
-  const [isCopied, setIsCopied] = useState<boolean>(false);
-
-  const copyToClipboard = (text: string) => {
-    navigator.clipboard
-      .writeText(text)
-      .then(() => {
-        setIsCopied(true);
-        setTimeout(() => setIsCopied(false), 2000);
-      })
-      .catch(() => {
-        setIsCopied(false);
-      });
-  };
-
-  return [isCopied, copyToClipboard];
-};
-
-export default useCopyToClipboard;
